Backport: coprocess_python: make sure that middleware reload and driver initialization use the OS thread (#970)
To avoid refactoring the rest of the drivers I've used a channel to signal when the initialization is done.
The specific cgo calls are wrapped as a goroutine that uses the appropriate runtime.LockOSThread and runtime.UnlockOSThread methods.
Added proper thread state handlers for safety.
