+++ coprocess_bundle.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/TykTechnologies/logrus"
 	"github.com/TykTechnologies/goverify"
+	"github.com/TykTechnologies/logrus"
 	"github.com/TykTechnologies/tykcommon"
 
 	"archive/zip"
@@ -391,17 +391,4 @@ func bundleError(spec *APISpec, err error, message string) {
 		"api_id":      spec.APIDefinition.APIID,
 		"path":        "-",
 	}).Error(message, ": ", err)
-}
-
-// getBundlePaths will return an array of the available bundle directories:
-func getBundlePaths() []string {
-	directories := make([]string, 0)
-	bundles, _ := ioutil.ReadDir(tykBundlePath)
-	for _, f := range bundles {
-		if f.IsDir() {
-			fullPath := filepath.Join(tykBundlePath, f.Name())
-			directories = append(directories, fullPath)
-		}
-	}
-	return directories
 }
